+++ resources/release-notes.md
@@ -1,3 +1,7 @@
+## App
+
+- (HelioSwarm) Add support for visualizing Range events with field
+
 ## Software
 
 - (All) Add option to specify PNG resolution when saving figures